move special proposal structs to own header
